+++ Public/Get-SeElement.ps1
@@ -98,10 +98,9 @@ function Get-SeElement {
             $AdditionalAttributes | ForEach-Object {
                 if (!$MyAttributes.Contains($_)) { $MyAttributes.Add($_) }
             }
-            $Attributes = [String[]]$MyAttributes
         }
 
-        if ($Attributes) {
+        if ($MyAttributes.Count -gt 0) {
             
             if ($GetAllAttributes) {
                 Foreach ($Item in $Output) {
@@ -119,7 +118,7 @@ function Get-SeElement {
                 foreach ($Item in $Output) {
                     $AttArray = [System.Collections.Generic.Dictionary[String, String]]::new()
                  
-                    foreach ($att in $Attributes) {
+                    foreach ($att in $MyAttributes) {
                         $attValue = $Item.GetAttribute($att)
                         if ($attValue -ne "") {
                             $AttArray.Add($att, $Item.GetAttribute($att))